requirements.txt
numpy
pandas
google-genai
python-dotenv
